% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auc.R
\name{auc.cure}
\alias{auc.cure}
\title{AUC for mixture cure models}
\usage{
auc.cure(odds, risk, surv, delta)
}
\arguments{
\item{odds}{odds for cure part}

\item{risk}{risk for cox part}

\item{surv}{survival probability}

\item{delta}{censoring status}
}
\description{
AUC for mixture cure models
}

