% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cindex.R
\name{cindex}
\alias{cindex}
\title{C index for mixture cure models}
\usage{
cindex(risk, pi, time, delta, tau = NULL)
}
\arguments{
\item{risk}{the risk score for survival part}

\item{pi}{the uncured probability}

\item{time}{observation time}

\item{delta}{censoring status}

\item{tau}{boundry of time}
}
\description{
Nonparametrics estimators for C index
}
\examples{
fit <- coxph(Surv(time, status) ~  age + sex , lung, x = T)
risk <- fit$x \%*\% fit$coefficients
survConcordance(Surv(time, status) ~ risk, lung)
pi <- rep(1,fit$n)
c.index(risk, pi, lung$time, lung$status ==2)
fit <- coxph(Surv(stop, event) ~  rx + size , bladder, x = T)
risk <- fit$x \%*\% fit$coefficients
survConcordance(Surv(stop, event) ~ risk, bladder)
pi <- rep(1,fit$n)
c.index(risk, pi, bladder$stop, bladder$event)
}
