% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auc.R
\name{cutSenspe}
\alias{cutSenspe}
\title{Sensitivity and specificity for a cut point}
\usage{
cutSenspe(res, cut.point)
}
\arguments{
\item{cut.point}{cutpoint for sensitivity and specificity}

\item{results}{of sen.spe.cure() when cut.point = NULL}
}
\description{
Sensitivity and specificity for a cut point
}
