% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eva_cure_direct.R
\name{eva_cure_direct}
\alias{eva_cure_direct}
\title{Evaluate the direct estimator}
\usage{
eva_cure_direct(
  time,
  delta,
  X,
  beta,
  Z,
  b,
  surv,
  model,
  cutpoint,
  n_post = 500,
  baseline = TRUE,
  posterior = TRUE
)
}
\description{
Evaluate the direct estimator
}
