% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smcure1.R
\name{eva_smcure}
\alias{eva_smcure}
\title{Evaluate Mixture cure model fitted by smcure}
\usage{
eva_smcure(fit, model)
}
\arguments{
\item{fit}{an extended smcure object}

\item{model}{the type of survival model ("PH", "PO","Normal")}
}
\description{
Evaluate Mixture cure model fitted by smcure
}

