% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kindex.R
\name{G_est}
\alias{G_est}
\title{Estimate the kernel function G}
\usage{
G_est(F_est, r, u = seq(-10, 10, length = 2000))
}
\arguments{
\item{F_est}{CDF of error distribution}

\item{r}{risk difference}

\item{u}{the design sequence point}
}
\description{
Estimate the kernel function G
}

