% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kindex.R
\name{k.ind.est}
\alias{k.ind.est}
\title{K-index estimation (aft model)}
\usage{
k.ind.est(risk, pi, F_est)
}
\arguments{
\item{risk}{the risk score for survival part}

\item{pi}{the uncured probability}

\item{F_est}{estimated error distribution}
}
\description{
K-index estimation (aft model)
}
