% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kindex.R
\name{k.ind}
\alias{k.ind}
\title{K-index estimation}
\usage{
k.ind(risk, pi, model)
}
\arguments{
\item{risk}{the risk score for survival part}

\item{pi}{the uncured probability}

\item{model}{the type of survival model ("PH", "PO","Normal")}
}
\description{
K-index estimation
}
