% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kindex.R
\name{k.index}
\alias{k.index}
\title{K-index estimation with variance}
\usage{
k.index(beta, b, coef.var, X, Z, model)
}
\arguments{
\item{beta}{parameters for survival part}

\item{b}{parameters for cure part}

\item{coef.var}{estimated variance-covariance matrix for parameters}

\item{X}{covariates for survival part}

\item{Z}{covariates for cure part}

\item{model}{the type of survival model ("PH", "PO","Normal")}
}
\description{
K-index estimation with variance
}

