% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smcure1.R
\name{printsmcure1}
\alias{printsmcure1}
\title{Print smcure object with evaluation information}
\usage{
printsmcure1(x, Var = TRUE, ROC = TRUE, ...)
}
\description{
Refer printsmcure() for detail information
}

