% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auc.R
\name{sen.spe.cure}
\alias{sen.spe.cure}
\title{Sensitivity and Specificity for mixture cure models}
\usage{
sen.spe.cure(odds, risk, surv, delta, cut.point = NULL, verbose = T)
}
\arguments{
\item{odds}{odds for cure part}

\item{risk}{risk for cox part}

\item{surv}{survival probability}

\item{delta}{censoring status}

\item{cut.point}{cut point for sensitivity (if NULL report all)}

\item{verbose}{if TRUE, report AUC and draw ROC curve}
}
\description{
Sensitivity and Specificity for mixture cure models
}
