% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auc.R
\name{w.cure}
\alias{w.cure}
\title{Posterior of uncure probabilty}
\usage{
w.cure(pi, delta, surv)
}
\arguments{
\item{pi}{uncure probability}

\item{delta}{censoring status}

\item{surv}{survival probability}
}
\description{
Posterior of uncure probabilty
}
